requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pymupdf==1.23.8
pytesseract==0.3.10
spacy==3.7.2
chromadb
whoosh==2.7.4
networkx==3.2
requests==2.31.0
python-dateutil==2.8.2
regex==2023.12.25
aiofiles==23.2.1
openai==1.12.0
numpy==1.24.3
scikit-learn==1.3.2
PyPDF2==3.0.1
python-pptx==0.6.23
python-docx==1.1.0
reportlab==4.0.7
Pillow==10.1.0